requirements.txt
fastapi
uvicorn
PyMuPDF
langchain
openai
sentence-transformers
faiss-cpu
chromadb
pydantic
python-multipart
groq
llama-index
google-generativeai
python-dotenv
langchain-google-genai
